+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+EMCP.SessionStore.ETS.init()
 ExUnit.start()